# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sunrise Staff Housing Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ASSIGNMENTS
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        nationalId:
          type: string
        nationality:
          type: string
        address:
          type: string
        jobTitle:
          type: string
        level:
          type: string
        phone:
          type: string
        department:
          type: string
        status:
          type: string
        hireDate:
          type: string
        gender:
          type: string
        idImage:
          type: ["string", "null"]
      required:
        - id
        - propertyId
        - employeeId
        - firstName
        - lastName
        - nationalId
        - nationality
        - address
        - jobTitle
        - level
        - phone
        - department
        - status
        - hireDate
        - gender
...
    UpdateEmployeeBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        nationalId:
          type: string
        nationality:
          type: string
        address:
          type: string
        jobTitle:
          type: string
        level:
          type: string
        phone:
          type: string
        department:
          type: string
        status:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
        idImage:
          type: ["string", "null"]
...